Can have animated images in textbox, fixed minor issues

+++ Glyph.h
@@ -9,13 +9,16 @@
 struct Glyph
 {
 	Sprite* sprite = nullptr;
+	Animator* animator = nullptr; // never delete these
 	Vector2 position = Vector2(0, 0);
+	bool shouldDeleteSprite = true;
 	//float width = 0; Should we save width and height to avoid calculations?
 	// Maybe if the value is negative, recalculate, otherwise use what is there?
 
 	~Glyph()
 	{
-		delete_it(sprite);
+		if (shouldDeleteSprite)
+			delete_it(sprite);
 	}
 };
